Add the function call pattern to syntax
Add the pattern to name the function calls with parenthesis or dot and
parenthesis.

There is a full explanation about why, on the GitHub PR of this commit.
TL;DR Since function calls without the parenthesis raise a compile
warning and the formatter add the parenthesis by default, we should
support a rich experience for those who follow the language
recomendation.